+++ Reflection/src/utopia/reflection/component/swing/input/Slider.scala
@@ -339,24 +339,10 @@ class Slider[+A](range: AnyAnimation[A], targetKnobDiameter: Double, targetWidth
 		if (state.isInFocus) base + 1 else base
 	}
 	
-	private def currentKnobColor =
-	{
+	private def currentKnobColor = {
 		val base = knobColor(doubleValue)
 		val changes = colorChangeIterations
-		val changed =
-		{
-			if (changes > 0)
-			{
-				if (base.luminosity < 0.6)
-					Iterator.iterate(base) { _.lightened(1 + 0.4 * colorVariationIntensity) }
-						.drop(changes).next()
-				else
-					Iterator.iterate(base) { _.darkened(1 + 0.4 * colorVariationIntensity) }
-						.drop(changes).next()
-			}
-			else
-				base
-		}
+		val changed = if (changes > 0) base.highlightedBy(changes) else base
 		if (enabled)
 			changed
 		else